Added custom hooks
